refactor(front): add postEditor component
