+++ mylog/src/navigations/stack/FeedStackNavigator.tsx
@@ -3,18 +3,19 @@ import {
   StackNavigationOptions,
   createStackNavigator,
 } from '@react-navigation/stack';
+import type {LatLng} from 'react-native-maps';
 
 import FeedHomeScreen from '@/screens/feed/FeedHomeScreen';
 import FeedDetailScreen from '@/screens/feed/FeedDetailScreen';
+import EditPostScreen from '@/screens/feed/EditPostScreen';
 import FeedHomeHeaderLeft from '@/components/feed/FeedHomeHeaderLeft';
 import {feedNavigations} from '@/constants/navigations';
 import {colors} from '@/constants/colors';
-import EditPostScreen from '@/screens/feed/EditPostScreen';
 
 export type FeedStackParamList = {
   [feedNavigations.FEED_HOME]: undefined;
   [feedNavigations.FEED_DETAIL]: {id: number};
-  [feedNavigations.EDIT_POST]: {id: number};
+  [feedNavigations.EDIT_POST]: {location: LatLng};
 };
 
 const Stack = createStackNavigator<FeedStackParamList>();
