+++ mylog/src/screens/map/AddPostScreen.tsx
@@ -1,189 +1,19 @@
-import React, {useCallback, useLayoutEffect, useRef, useState} from 'react';
-import {
-  SafeAreaView,
-  ScrollView,
-  StyleSheet,
-  TextInput,
-  View,
-} from 'react-native';
+import React from 'react';
 import type {StackScreenProps} from '@react-navigation/stack';
-import Octicons from 'react-native-vector-icons/Octicons';
 
 import type {MapStackParamList} from '@/navigations/stack/MapStackNavigator';
-import InputField from '@/components/@common/InputField';
-import CustomKeyboardAvoidingView from '@/components/@keyboard/CustomKeyboardAvoidingView';
-import AddPostHeaderRight from '@/components/post/AddPostHeaderRight';
-import CustomButton from '@/components/@common/CustomButton';
-import MarkerSelector from '@/components/post/MarkerSelector';
-import ScoreInput from '@/components/post/ScoreInput';
-import ImageInput from '@/components/post/ImageInput';
-import PreviewImageList from '@/components/post/PreviewImageList';
-import DatePickerModal from '@/components/post/DatePickerModal';
-import {useCreatePost} from '@/hooks/queries/usePost';
-import useGetAddress from '@/hooks/common/useGetAddress';
-import useDatePicker from '@/hooks/common/useDatePicker';
-import useForm from '@/hooks/common/useForm';
-import usePermission from '@/hooks/common/usePermission';
-import useImageInput from '@/hooks/common/useImageInput';
-import {validateAddPost} from '@/utils/validate';
-import {getDateWithSeparator} from '@/utils/date';
+import PostEditor from '@/components/post/PostEditor';
 import {mapNavigations} from '@/constants/navigations';
-import {colors} from '@/constants/colors';
-import type {MarkerColor} from '@/types/domain';
 
 type AddPostScreenProps = StackScreenProps<
   MapStackParamList,
   typeof mapNavigations.ADD_POST
 >;
 
-function AddPostScreen({route, navigation}: AddPostScreenProps) {
+function AddPostScreen({route}: AddPostScreenProps) {
   const {location} = route.params;
-  const descriptionRef = useRef<TextInput | null>(null);
-  const [marker, setMarker] = useState<MarkerColor>('RED');
-  const [score, setScore] = useState(3);
-  const datePicker = useDatePicker(new Date());
-  const createPostMutation = useCreatePost();
-  const address = useGetAddress(location);
-  const imageInput = useImageInput();
-  const addPost = useForm({
-    initialValue: {title: '', description: ''},
-    validate: validateAddPost,
-  });
-  usePermission('PHOTO');
 
-  const handleSubmit = useCallback(() => {
-    createPostMutation.mutate(
-      {
-        latitude: location.latitude,
-        longitude: location.longitude,
-        color: marker,
-        title: addPost.values.title,
-        description: addPost.values.description,
-        date: datePicker.date,
-        address,
-        score,
-        imageUris: imageInput.imageUris,
-      },
-      {
-        onSuccess: () => navigation.goBack(),
-        onError: err => console.log(err.response?.data.message),
-      },
-    );
-  }, [
-    createPostMutation,
-    location,
-    marker,
-    address,
-    score,
-    addPost.values,
-    datePicker.date,
-    imageInput.imageUris,
-    navigation,
-  ]);
-
-  const handleSelectMarker = (name: MarkerColor) => {
-    setMarker(name);
-  };
-
-  const handleChangeScore = (value: number) => {
-    setScore(value);
-  };
-
-  useLayoutEffect(() => {
-    navigation.setOptions({
-      headerRight: () => AddPostHeaderRight(handleSubmit, addPost.hasErrors),
-    });
-  }, [handleSubmit, navigation, addPost.hasErrors]);
-
-  return (
-    <SafeAreaView style={styles.container}>
-      <CustomKeyboardAvoidingView>
-        <ScrollView
-          style={styles.contentContainer}
-          scrollIndicatorInsets={{right: 1}}>
-          <View style={styles.inputContainer}>
-            <InputField
-              value={address}
-              disabled={true}
-              icon={
-                <Octicons name="location" size={16} color={colors.GRAY_500} />
-              }
-            />
-            <CustomButton
-              variant="outlined"
-              size="large"
-              label={
-                datePicker.isPicked
-                  ? `${getDateWithSeparator(datePicker.date, '. ')}`
-                  : '날짜 선택'
-              }
-              onPress={datePicker.showModal}
-            />
-            <InputField
-              {...addPost.getTextInputProps('title')}
-              error={addPost.errors.title}
-              touched={addPost.touched.title}
-              placeholder="제목을 입력하세요."
-              maxLength={30}
-              returnKeyType="next"
-              blurOnSubmit={false}
-              onSubmitEditing={() => {
-                descriptionRef.current?.focus();
-              }}
-            />
-            <InputField
-              {...addPost.getTextInputProps('description')}
-              error={addPost.errors.description}
-              touched={addPost.touched.description}
-              ref={descriptionRef}
-              placeholder="기록하고 싶은 내용을 입력하세요. (선택)"
-              maxLength={3000}
-              returnKeyType="next"
-              multiline
-            />
-            <MarkerSelector
-              marker={marker}
-              onPressMarker={handleSelectMarker}
-            />
-            <ScoreInput score={score} onChangeScore={handleChangeScore} />
-            <View style={styles.imagesViewer}>
-              <ImageInput onChange={imageInput.handleChange} />
-              <PreviewImageList
-                imageUris={imageInput.imageUris}
-                onDelete={imageInput.delete}
-                onChangeOrder={imageInput.changeOrder}
-                showOption
-              />
-            </View>
-            <DatePickerModal
-              date={datePicker.date}
-              isVisible={datePicker.isVisible}
-              onChangeDate={datePicker.handleChange}
-              onConfirmDate={datePicker.handleConfirm}
-            />
-          </View>
-        </ScrollView>
-      </CustomKeyboardAvoidingView>
-    </SafeAreaView>
-  );
+  return <PostEditor location={location} />;
 }
-
-const styles = StyleSheet.create({
-  container: {
-    flex: 1,
-  },
-  contentContainer: {
-    flex: 1,
-    padding: 20,
-    marginBottom: 10,
-  },
-  inputContainer: {
-    gap: 20,
-    marginBottom: 20,
-  },
-  imagesViewer: {
-    flexDirection: 'row',
-  },
-});
 
 export default AddPostScreen;
